Added class QuadraticLikeCost.
In this way we can have a common ancestro between a generic quadratic cost, a L2Norm cost and a linear cost without the need of hiding QuadraticCost specialization methods.
